Cygwin: is_console_app(): handle app execution aliases
In 0a9ee3ea23 (Allow executing Windows Store's "app execution aliases",
2021-03-12), I introduced support for calling Microsoft Store
applications.

However, it was reported several times (first in
https://inbox.sourceware.org/cygwin/CAAM_cieBo_M76sqZMGgF+tXxswvT=jUHL_pShff+aRv9P1Eiug@mail.gmail.com
and then also in
https://github.com/msys2/MSYS2-packages/issues/1943#issuecomment-3467583078)
that there is something amiss: The standard handles are not working as
expected, as they are not connected to the terminal at all (and hence
the application seems to "hang").

The culprit is the `is_console_app()` function which assumes that it can
simply open the first few bytes of the `.exe` file to read the PE header
in order to determine whether it is a console application or not.

For app execution aliases, already creating a regular file handle for
reading will fail. Let's introduce some special handling for the exact
error code returned in those instances, and try to read the symlink
target instead (taking advantage of the code I introduced in 0631c6644e
(Cygwin: Treat Windows Store's "app execution aliases" as symbolic
links, 2021-03-22) to treat app execution aliases like symbolic links).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>